feat: Add dependency tracking to Spec options
This commit adds the `dependsOn` and `dependsOnValue` properties to the `Spec` type in the `src/types/spec.ts` file. These properties allow for dependency tracking between options in the product form. When an option depends on another option, it will be conditionally displayed based on the selected value of the dependency option.

The `dependsOn` property specifies the name of the option that the current option depends on, while the `dependsOnValue` property specifies the value of the dependency option that triggers the display of the current option.

This change enables more dynamic and interactive behavior in the product form.